requirements.txt
Adafruit-Blinka==8.31.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel-spi==1.0.9
adafruit-circuitpython-pixelbuf==2.0.4
adafruit-circuitpython-requests==2.0.4
adafruit-circuitpython-typing==1.9.6
Adafruit-PlatformDetect==3.59.0
Adafruit-PureIO==1.1.11
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pillow==10.2.0
PyAudio==0.2.14
pyftdi==0.55.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
six==1.16.0
typing_extensions==4.9.0
